+++ src/Exception/Kafka/KafkaProducerException.php
@@ -8,6 +8,9 @@ final class KafkaProducerException extends KafkaException
 {
     public static function flushFailed(int $errorCode): self
     {
-        return new self(\sprintf('Flush failed: %s (%d)', rd_kafka_err2str($errorCode), $errorCode));
+        return new self(
+            \sprintf('Flush failed: %s (%d)', rd_kafka_err2str($errorCode), $errorCode),
+            $errorCode,
+        );
     }
 }